requirements.txt
selenium
python-anticaptcha
requests
selenium-wire
pytz
python-dateutil